feat(testing): add comprehensive unit and integration tests
Adds a comprehensive suite of unit and integration tests for core application logic, focusing on hooks and the TimerControls component.

- Creates new test files for `TimerControls`, `useCalorieTracker`, `useWorkoutSessionManager`, and `useBluetoothHRM`.
- Consolidates existing tests for `TimerControls` into a single file.
- Adds tests for `TIMER_CONFIG`, `TIMER_COMMAND`, and `SET_MODE` messages in `TimerControls`.
- Adds tests for the `useCalorieTracker` hook to verify calorie calculation and reset functionality.
- Adds tests for the `useWorkoutSessionManager` hook to verify the entire workout lifecycle.
- Adds regression tests for the `useBluetoothHRM` hook's reconnection logic.
- Ensures existing tests for `useLocalWorkoutBuffer` are sufficient.
- Fixes linting errors in new and existing test files.
