+++ app/client/control/components/TimerControls.test.tsx
@@ -128,4 +128,40 @@ describe('TimerControls', () => {
     render(<TimerControls />)
     expect(screen.getByTestId('timer-stopped')).toBeInTheDocument()
   })
+
+  it('should revert optimistic state when server state changes', async () => {
+    const initialTimerData = {
+      isRunning: false,
+      currentPhase: 'WORK',
+      mode: 'TABATA',
+    }
+
+    mockUseWebSocket.mockReturnValue({
+      timerData: initialTimerData,
+      sendData: mockSendData,
+      connectionStatus: 'Connected',
+    })
+
+    const { rerender } = render(<TimerControls />)
+
+    // User clicks START, UI optimistically updates to 'Running'
+    fireEvent.click(screen.getByTestId('start-timer-button'))
+    expect(screen.getByTestId('timer-running')).toBeInTheDocument()
+
+    // Server sends a delayed response indicating the timer is NOT running
+    act(() => {
+      mockUseWebSocket.mockReturnValue({
+        timerData: { ...initialTimerData, isRunning: false },
+        sendData: mockSendData,
+        connectionStatus: 'Connected',
+      })
+    })
+
+    rerender(<TimerControls />)
+
+    // The UI should revert to the server's state
+    await waitFor(() => {
+      expect(screen.getByTestId('timer-stopped')).toBeInTheDocument()
+    })
+  })
 })
